requirements.txt
appdirs==1.4.4
audioread==2.1.9
autosubsync==1.0.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
decorator==4.4.2
ffprobe==0.5
google-api-core==2.7.3
google-api-python-client==2.47.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
googleapis-common-protos==1.56.0
httplib2==0.20.4
idna==3.3
imageio==2.19.2
imageio-ffmpeg==0.4.7
joblib==1.1.0
llvmlite==0.38.0
moviepy==1.0.3
numba==0.55.1
numpy==1.21.6
oauth2client==4.1.3
oauthlib==3.2.0
packaging==21.3
pedalboard==0.5.4
Pillow==9.1.1
pooch==1.6.0
proglog==0.1.10
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydub==0.25.1
pygame==2.1.2
pylrc==0.1.2
pyparsing==3.0.8
pysndfx==0.3.6
PySoundFile==0.9.0.post1
requests==2.27.1
requests-oauthlib==1.3.1
resampy==0.2.2
rsa==4.8
scikit-learn==1.1.0
scipy==1.8.1
six==1.16.0
SoundFile==0.10.3.post1
sox==1.4.1
threadpoolctl==3.1.0
tqdm==4.64.0
uritemplate==4.1.1
urllib3==1.26.9
